+++ src/addressablestools/__init__.py
@@ -9,7 +9,7 @@ from addressablestools.models import (
 )
 from addressablestools.parser import parse, parse_binary, parse_json
 
-__version__ = "1.1.0"
+__version__ = "1.1.1"
 
 __all__ = [
     "AssetBundleRequestOptions",